portfolio_generator/requirements.txt
openai>=1.1.0
python-dotenv>=1.0.0
requests>=2.25.0
asyncio>=3.4.3
google-cloud-firestore>=2.10.0
google-cloud-storage>=2.14.0
opencv-python>=4.8.0
pillow>=10.0.0
tiktoken>=0.5.1  # optional, for token counting
numpy>=1.24.0
tabulate>=0.9.0
rich>=12.0.0
docopt>=0.6.2
